std::rand: move normal and exponential to their own file.
